+++ crates/cli/src/introspection/sampling.rs
@@ -1,7 +1,7 @@
-use std::collections::BTreeMap;
+use std::collections::{BTreeMap, HashSet};
 
 use super::type_unification::{
-    unify_object_types, unify_schema, unify_type, TypeUnificationContext, TypeUnificationResult,
+    unify_object_types, unify_type, TypeUnificationContext, TypeUnificationResult,
 };
 use configuration::{
     schema::{self, Type},
@@ -22,21 +22,21 @@ type ObjectType = WithName<schema::ObjectType>;
 pub async fn sample_schema_from_db(
     sample_size: u32,
     config: &MongoConfig,
-) -> anyhow::Result<Schema> {
-    let mut schema = Schema {
-        collections: BTreeMap::new(),
-        object_types: BTreeMap::new(),
-    };
+    existing_schemas: &HashSet<std::string::String>,
+) -> anyhow::Result<BTreeMap<std::string::String, Schema>> {
+    let mut schemas = BTreeMap::new();
     let db = config.client.database(&config.database);
     let mut collections_cursor = db.list_collections(None, None).await?;
 
     while let Some(collection_spec) = collections_cursor.try_next().await? {
         let collection_name = collection_spec.name;
-        let collection_schema =
-            sample_schema_from_collection(&collection_name, sample_size, config).await?;
-        schema = unify_schema(schema, collection_schema);
+        if !existing_schemas.contains(&collection_name) {
+            let collection_schema =
+                sample_schema_from_collection(&collection_name, sample_size, config).await?;
+            schemas.insert(collection_name, collection_schema);
+        }
     }
-    Ok(schema)
+    Ok(schemas)
 }
 
 async fn sample_schema_from_collection(