+++ crates/configuration/src/directory.rs
@@ -3,15 +3,15 @@ use futures::stream::TryStreamExt as _;
 use itertools::Itertools as _;
 use serde::{Deserialize, Serialize};
 use std::{
-    collections::BTreeMap,
+    collections::{BTreeMap, HashSet},
     path::{Path, PathBuf},
 };
 use tokio::fs;
 use tokio_stream::wrappers::ReadDirStream;
 
-use crate::{with_name::WithName, Configuration};
+use crate::{with_name::WithName, Configuration, Schema};
 
-pub const SCHEMA_FILENAME: &str = "schema";
+pub const SCHEMA_DIRNAME: &str = "schema";
 pub const NATIVE_QUERIES_DIRNAME: &str = "native_queries";
 
 pub const CONFIGURATION_EXTENSIONS: [(&str, FileFormat); 3] =
@@ -33,7 +33,10 @@ pub async fn read_directory(
 ) -> anyhow::Result<Configuration> {
     let dir = configuration_dir.as_ref();
 
-    let schema = parse_json_or_yaml(dir, SCHEMA_FILENAME).await?;
+    let schemas = read_subdir_configs(&dir.join(SCHEMA_DIRNAME))
+        .await?
+        .unwrap_or_default();
+    let schema = schemas.into_values().fold(Schema::default(), Schema::merge);
 
     let native_queries = read_subdir_configs(&dir.join(NATIVE_QUERIES_DIRNAME))
         .await?
@@ -100,41 +103,6 @@ where
     }
 }
 
-/// Given a base name, like "connection", looks for files of the form "connection.json",
-/// "connection.yaml", etc; reads the file; and parses it according to its extension.
-async fn parse_json_or_yaml<T>(configuration_dir: &Path, basename: &str) -> anyhow::Result<T>
-where
-    for<'a> T: Deserialize<'a>,
-{
-    let (path, format) = find_file(configuration_dir, basename).await?;
-    parse_config_file(path, format).await
-}
-
-/// Given a base name, like "connection", looks for files of the form "connection.json",
-/// "connection.yaml", etc, and returns the found path with its file format.
-async fn find_file(
-    configuration_dir: &Path,
-    basename: &str,
-) -> anyhow::Result<(PathBuf, FileFormat)> {
-    for (extension, format) in CONFIGURATION_EXTENSIONS {
-        let path = configuration_dir.join(format!("{basename}.{extension}"));
-        if fs::try_exists(&path).await? {
-            return Ok((path, format));
-        }
-    }
-
-    Err(anyhow!(
-        "could not find file, {:?}",
-        configuration_dir.join(format!(
-            "{basename}.{{{}}}",
-            CONFIGURATION_EXTENSIONS
-                .into_iter()
-                .map(|(ext, _)| ext)
-                .join(",")
-        ))
-    ))
-}
-
 async fn parse_config_file<T>(path: impl AsRef<Path>, format: FileFormat) -> anyhow::Result<T>
 where
     for<'a> T: Deserialize<'a>,
@@ -149,12 +117,31 @@ where
     Ok(value)
 }
 
-/// Currently only writes `schema.json`
-pub async fn write_directory(
+async fn write_subdir_configs<T>(
+    subdir: &Path,
+    configs: impl IntoIterator<Item = (String, T)>,
+) -> anyhow::Result<()>
+where
+    T: Serialize,
+{
+    if !(fs::try_exists(subdir).await?) {
+        fs::create_dir(subdir).await?;
+    }
+
+    for (name, config) in configs {
+        let with_name: WithName<T> = (name.clone(), config).into();
+        write_file(subdir, &name, &with_name).await?;
+    }
+
+    Ok(())
+}
+
+pub async fn write_schema_directory(
     configuration_dir: impl AsRef<Path>,
-    configuration: &Configuration,
+    schemas: impl IntoIterator<Item = (String, Schema)>,
 ) -> anyhow::Result<()> {
-    write_file(configuration_dir, SCHEMA_FILENAME, &configuration.schema).await
+    let subdir = configuration_dir.as_ref().join(SCHEMA_DIRNAME);
+    write_subdir_configs(&subdir, schemas).await
 }
 
 fn default_file_path(configuration_dir: impl AsRef<Path>, basename: &str) -> PathBuf {
@@ -175,4 +162,17 @@ where
     fs::write(path.clone(), bytes)
         .await
         .with_context(|| format!("error writing {:?}", path))
+}
+
+pub async fn list_existing_schemas(
+    configuration_dir: impl AsRef<Path>,
+) -> anyhow::Result<HashSet<String>> {
+    let dir = configuration_dir.as_ref();
+
+    // TODO: we don't really need to read and parse all the schema files here, just get their names.
+    let schemas = read_subdir_configs::<Schema>(&dir.join(SCHEMA_DIRNAME))
+        .await?
+        .unwrap_or_default();
+
+    Ok(schemas.into_keys().collect())
 }