+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 This changelog documents the changes between release versions.
 
 ## [Unreleased]
-Changes to be included in the next upcoming release
+- Use separate schema files for each collection
+- Don't sample from collections that already have a schema
 
 ## [0.0.2] - 2024-03-26
 - Rename CLI plugin to ndc-mongodb ([PR #13](https://github.com/hasura/ndc-mongodb/pull/13))